pyproject.toml
[project]
name = "scholaragent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "python-multipart",
    "httpx",
    "sqlalchemy>=2.0",
    "psycopg2-binary",
    "alembic",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
